Added Teams, Improved Interactions
Added the ability for players to use their num keys to switch to different teams, based on this we are able to recolor characters, prevent friendly fire, and constrain unfreezing to players on the same team.

Also improved the code for interactions making it a slight bit more general purpose, still needs some more love to enable picking up items and interacting with climbzones.