Interaction Code Refactor
Refactored the interaction code to make it much simpler to read. I think there are still a number of improvements that can be made to further simplify the code, encapsulate all UI code to better align that with player abilities, and opportunities to make the code easier to extend in the case where we add further types of interactables.  A few known bugs exist with the current build, mostly with UI and the ability for players to tag already frozen players.